% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertising.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Data frame representing information about advertisements}
\format{
A tibble with 200 rows and 4 variables:

\describe{
}
}
\source{
\url{https://www.statlearning.com/}
}
\usage{
advertising
}
\description{
This simulated data is from Introduction to Statistical Learning (ISLR) book.
}
\keyword{datasets}
