% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlanta_commute.R
\docType{data}
\name{atlanta_commute}
\alias{atlanta_commute}
\title{Random sample of commuters from Atlanta}
\format{
A tibble with 500 rows and 5 variables:

\describe{
}
}
\source{
Unknown
}
\usage{
atlanta_commute
}
\description{
Random sample of commuters from Atlanta
}
\keyword{datasets}
