% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_hours.R
\docType{data}
\name{bike_hours}
\alias{bike_hours}
\title{DC Bike Share data}
\format{
A tibble with 17,379  rows and 17 variables:

\describe{
}
}
\source{
\url{http://dx.doi.org/10.1007/s13748-013-0040-3}
}
\usage{
bike_hours
}
\description{
Data contains info on bike rentals in DC in 2011 and 2012.
}
\keyword{datasets}
