% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\docType{data}
\name{county}
\alias{county}
\title{United States Counties}
\format{
A tibble with 3,142 rows and 15 variables:

\describe{
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=county}
}
\usage{
county
}
\description{
Data for 3142 counties in the United States
}
\keyword{datasets}
