% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_side_barplot.R
\name{create_side_barplot}
\alias{create_side_barplot}
\title{Create a barplot}
\usage{
create_side_barplot(df, cat_var)
}
\arguments{
\item{df}{tibble}

\item{cat_var}{categorical variable}
}
\value{
side, sorted, barplot
}
\description{
This function creates a 'sorted' and flipped barplot
for a categorical variable in dataframe. It assumes that
the first argument is a tibble and the second argument
is a categorical variable (specifically a factor type)
from that tibble
}
