% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure_delays.R
\docType{data}
\name{departure_delays}
\alias{departure_delays}
\title{Departure delays in Syracuse Hancock International Airport in 2019}
\format{
A tibble with 3,161 rows and 6 variables:

\describe{
}
}
\source{
\url{https://www.transtats.bts.gov/}
}
\usage{
departure_delays
}
\description{
Three major airlines: United Airlines, Delta Airlines and American Airlines
}
\keyword{datasets}
