% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df1_county.R
\docType{data}
\name{df1_county}
\alias{df1_county}
\title{Part of County data without 'metro' variable}
\format{
A tibble with 3,142  rows and 14 variables:

\describe{
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=county}
}
\usage{
df1_county
}
\description{
Part of County data without 'metro' variable
}
\keyword{datasets}
