% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2_county.R
\docType{data}
\name{df2_county}
\alias{df2_county}
\title{Part of County data with only 'metro' variable and no missing row}
\format{
A tibble with 3,139  rows and 1 variable:

\describe{
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=county}
}
\usage{
df2_county
}
\description{
Part of County data with only 'metro' variable and no missing row
}
\keyword{datasets}
