% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecommerce.R
\docType{data}
\name{ecommerce}
\alias{ecommerce}
\title{Simulated e-commerce A/B testing data}
\format{
A tibble with 787 rows and 2 variables:

\describe{
}
}
\source{
Simulated by Tural Sadigov.
}
\usage{
ecommerce
}
\description{
Simulated e-commerce A/B testing data
}
\keyword{datasets}
