% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\docType{data}
\name{emails}
\alias{emails}
\title{Data frame representing information about a collection of emails}
\format{
A tibble with 3,921 rows and 121 variables:

\describe{
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=email}
}
\usage{
emails
}
\description{
These data represent incoming emails for the first three months of 2012 for an email account (see Source).
}
\keyword{datasets}
