% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{In-hospital deaths from myocardial infarction with ST elevation (STEMI)
9 hospitals in the Providence Health system in Oregon.}
\format{
A tibble with 13 rows and 4 variables:

\describe{
}
}
\source{
\url{https://stat2-critical-look.netlify.app/the-bootstrap-method.html}
}
\usage{
hospital
}
\description{
In-hospital deaths from myocardial infarction with ST elevation (STEMI)
9 hospitals in the Providence Health system in Oregon.
}
\keyword{datasets}
