% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\docType{data}
\name{mice}
\alias{mice}
\title{Mice experiment data}
\format{
A tibble with 16 rows and 2 variables:

\describe{
}
}
\source{
Bradley Efron and Robert J. Tibshirani, 1998
}
\usage{
mice
}
\description{
Mice experiment data
}
\keyword{datasets}
