% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_steps.R
\name{ml_steps}
\alias{ml_steps}
\title{List Machine Learning Steps}
\usage{
ml_steps()
}
\value{
list of steps
}
\description{
This function lists the machine learning steps/pipeline
as we use it in MATH 254. Use it as a guide in your modeling
process
}
