% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist_sample.R
\docType{data}
\name{mnist_sample}
\alias{mnist_sample}
\title{Sample from famous MNIST data}
\format{
A tibble with 5,000  rows and 785 variables:

\describe{
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_sample
}
\description{
Data contains simple random sample of 5000 images that put into rows. The first column is the label.
}
\keyword{datasets}
