% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_salaries.R
\docType{data}
\name{nba_salaries}
\alias{nba_salaries}
\title{NBA salaries in 2019}
\format{
A tibble with 488 rows and 1 variables:

\describe{
}
}
\source{
Unknown
}
\usage{
nba_salaries
}
\description{
NBA salaries in 2019
}
\keyword{datasets}
