% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_births.R
\docType{data}
\name{nc_births}
\alias{nc_births}
\title{North Carolina births, 1000 cases}
\format{
A tibble with 1000 rows and 13 variables:

\describe{
In 2004, the state of North Carolina released to the public a large data set containing information on births recorded in this state. This data set has been of interest to medical researchers who are studying the relation between habits and practices of expectant mothers and the birth of their children. This is a random sample of 1,000 cases from this data set.
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=ncbirths}
}
\usage{
nc_births
}
\description{
Data for 3142 counties in the United States
}
\keyword{datasets}
