% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats2data_parallel.R
\name{plot.stats2data_parallel}
\alias{plot.stats2data_parallel}
\title{Plot RMSE distributions for a parallel result}
\usage{
\method{plot}{stats2data_parallel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stats2data_parallel}.}

\item{...}{Currently unused.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object, returned invisibly.
}
\description{
Visualises between-run variability and deviation-from-target RMSE
distributions as box-and-jitter plots, faceted by metric.
}
\examples{
\dontrun{
res <- parallel_optim(FUN = optim_aov, args = list(...), runs = 20)
plot(res)
}

}
