% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats2data_vec.R
\name{print.stats2data_vec}
\alias{print.stats2data_vec}
\title{Print a stats2data Descriptives result}
\usage{
\method{print}{stats2data_vec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stats2data_vec}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a stats2data Descriptives result
}
