% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats2data_aov.R
\name{print.summary.stats2data_aov}
\alias{print.summary.stats2data_aov}
\title{Print a stats2data ANOVA summary}
\usage{
\method{print}{summary.stats2data_aov}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.stats2data_aov}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a stats2data ANOVA summary
}
