% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats2data_mlr.R
\name{print.summary.stats2data_mlr}
\alias{print.summary.stats2data_mlr}
\title{Print a stats2data MLR summary}
\usage{
\method{print}{summary.stats2data_mlr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.stats2data_mlr}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a stats2data MLR summary
}
