% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{UK Smoking Data}
\format{
A tibble with 1,691 rows and 12 variables:

\describe{
Survey data on smoking habits from the UK. The data set can be used for analyzing the demographic characteristics of smokers and types of tobacco consumed.
}
}
\source{
\url{https://www.openintro.org/data/index.php?data=smoking}
}
\usage{
smoking
}
\description{
Data for 3142 counties in the United States
}
\keyword{datasets}
