% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{stats2data-package}
\alias{stats2data}
\alias{stats2data-package}
\title{stats2data: Data Simulation from Summary Statistics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We introduce stats2data: A Nonparametric Approach to Data Simulation from Summary Statistics. The primary scope of this algorithmic framework is to simulate complete raw datasets using only summary statistics, giving researchers a way to generate plausible sample data when original data are unavailable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastian-lortz/stats2data}
  \item \url{https://sebastian-lortz.github.io/stats2data/}
  \item Report bugs at \url{https://github.com/sebastian-lortz/stats2data/issues}
}

}
\author{
\strong{Maintainer}: Sebastian A. J. Lortz \email{s.a.j.lortz@gmail.com} (\href{https://orcid.org/0009-0007-0281-7996}{ORCID})

}
