% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_stats2data_parallel.R
\name{summary.stats2data_parallel}
\alias{summary.stats2data_parallel}
\title{Summarize a stats2data parallel result}
\usage{
\method{summary}{stats2data_parallel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{stats2data_parallel}.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{summary.stats2data_parallel}, printed by
  \code{\link{print.summary.stats2data_parallel}}.
}
\description{
Summarize a stats2data parallel result
}
