% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees}
\alias{trees}
\title{Data frame representing information about a measurements on trees}
\format{
A tibble with 70 rows and 2 variables:

\describe{
}
}
\source{
Unknown
}
\usage{
trees
}
\description{
Data frame representing information about a measurements on trees
}
\keyword{datasets}
