% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_side_barplot.R
\name{create_side_barplot}
\alias{create_side_barplot}
\title{Creates a 'sorted' and flipped barplot for a categorical variable in dataframe.}
\usage{
create_side_barplot(df, cat_var)
}
\arguments{
\item{df}{tibble}

\item{cat_var}{categorical variable}
}
\value{
side, sorted, barplot
}
\description{
Creates a 'sorted' and flipped barplot for a categorical variable in dataframe.
}
